refactor/vendor: add vendor route
